requirements.txt
python-twitter==3.5
importlib==1.0.4
ipython==8.2.0
jupyterlab==3.3.4
jupyter-contrib-nbextensions==0.5.1
DateTime==4.4
selenium==4.1.5
requests==2.27.1